requirements.txt
requests==2.13.0
aiohttp==1.2.0
appdirs==1.4.0
async-timeout==1.1.0
beautifulsoup4==4.5.3
cchardet==1.1.2
chardet==2.3.0
click==6.7
Flask==0.12
guess-language-spirit==0.5.2
itsdangerous==0.24
Jinja2==2.9.5
langdetect==1.0.7
lxml==3.7.2
MarkupSafe==0.23
multidict==2.1.4
packaging==16.8
pyparsing==2.1.10
PySocks==1.6.6
six==1.10.0
SQLAlchemy==1.1.5
translate==2.0.0
Werkzeug==0.11.15
yarl==0.9.1
